Move file list to own component
